pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"


[project]
name = "CoCoPyE"
version = "0.1.0"
authors = [
    { name = "Niklas Birth", email = "birth@posteo.de" },
]
description = "feature-based prediction of genome quality indices"
readme = "README.md"
requires-python = ">=3.8"
#keywords = ["one", "two"]
license = {text = "GPL-3.0-or-later"}
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "numpy>=1.23",
    "numba>=0.56.4",
    "biopython==1.81",
    "scikit-learn==1.3.1",
    "pandas",
    "packaging",
    "argparse~=1.4.0",
    "appdirs~=1.4.4",
    "tomlkit~=0.11.6",
    "requests~=2.31.0",
    "tqdm~=4.65.0",
    "numba-progress==1.0.0",
]

[project.optional-dependencies]
web = [
    "celery[redis]~=5.3.1",
    "uvicorn~=0.22.0",
    "fastapi~=0.99.1",
    "python-multipart~=0.0.6",
    "websockets~=11.0.3",
    "werkzeug",
    "Jinja2~=3.1.2"
]

[project.scripts]
cocopye = "cli:main"

# ... other project metadata fields as specified in:
#     https://packaging.python.org/en/latest/specifications/declaring-project-metadata/


[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]


[tool.mypy]
files = "src"
plugins = "numpy.typing.mypy_plugin"

strict = true
disallow_any_explicit = true
disallow_any_generics = true

[[tool.mypy.overrides]]
module = ["tomlkit", "Bio", "_io", "numba", "numba.typed"]
ignore_missing_imports = true